mlops_pipeline/requirements.txt
# Core Data Science
numpy<2.0.0
pandas>=1.5.0
scikit-learn>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Deep Learning
tensorflow>=2.11.0

# Cloud & Infrastructure
google-cloud-bigquery>=3.10.0
google-cloud-aiplatform>=1.25.0
google-cloud-storage>=2.9.0
python-dotenv>=1.0.0
pyarrow>=10.0.0
db-dtypes>=1.0.0
hydra-core>=1.3.2
typer[all]
rich
